requirements.txt
httpx
requests
python-dotenv
beautifulsoup4
langchain
langchain-google-genai
langchain-community
langchain-core
tiktoken
pytest
openai
pillow
streamlit